Use neoterm with vim-test
neovim mode with vim-test will leave in insert mode, meaning that
pressing enter will close the session - and also give focus to that
window.

Using neoterm will run the terminal without yielding focus to the test
being run, and is also run in normal mode, so you can scroll through the
test results more easily.

This also adds a custom keybinding to close the neoterm session. This is
by hitting leader-rr, which is the opposite of running one test with
leader-r.